requirements.txt
# Core dependencies
opencv-python>=4.8.0
numpy>=1.24.0
pyserial>=3.5
ultralytics>=8.0.0  # For YOLOv8

# Additional dependencies for better performance
torch>=2.0.0  # Required for YOLOv8
torchvision>=0.15.0 